formulaire de vente complet

+++ src/main/java/fr/eni/encheres/dal/CategorieDAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface CategorieDAO {
 
-    Categorie read(int no_categorie);
+    Categorie read(Long no_categorie);
     List<Categorie> findAll();
 
     void create(Categorie categorie) ;
